clean: separate impl From<reqwest::Response>
