+++ rust-connectors/sources/http/src/formatter.rs
@@ -1,5 +1,8 @@
 //! Output Record Formatting
 
+// Implements TryFrom<reqwest::Response>
+mod from_reqwest;
+
 #[derive(thiserror::Error, Debug)]
 pub enum HttpRecordError {}
 
@@ -26,7 +29,9 @@ impl HttpResponseRecord {
         let status_line: Vec<String> = vec![
             self.version.to_owned(),
             self.status_code.to_string(),
-            self.status_string.to_owned().unwrap_or_else(|| "".to_string()),
+            self.status_string
+                .to_owned()
+                .unwrap_or_else(|| "".to_string()),
         ];
         record_out_parts.push(status_line.join(" "));
 
@@ -51,57 +56,6 @@ impl HttpResponseRecord {
     }
 }
 
-// Reqwest Response TryFrom helper impl.
-fn version(version: &reqwest::Version) -> String {
-    format!("{:?}", version)
-}
-
-// Reqwest Response TryFrom helper impl.
-fn status(status: &reqwest::StatusCode) -> (u16, Option<String>) {
-    let status_code = status.as_u16();
-
-    let status_string =
-        status.canonical_reason().map(|s| s.to_string());
-
-    (status_code, status_string)
-}
-
-// Reqwest Response TryFrom helper impl.
-fn headers(hdr_map: &reqwest::header::HeaderMap) -> Vec<HttpHeader> {
-    let mut hdr_vec = Vec::with_capacity(hdr_map.len());
-
-    for (hdr_key, hdr_val) in hdr_map.iter() {
-        let hdr_value = match hdr_val.to_str() {
-            Ok(v) => v.to_owned().to_string(),
-            // Should we warn / error / spend effort on opportunistic lossy header?
-            Err(_) => String::from(""),
-        };
-
-        hdr_vec.push(HttpHeader {
-            name: hdr_key.to_owned().to_string(),
-            value: hdr_value,
-        });
-    }
-
-    hdr_vec
-}
-
-// Turn reqwest Response into HttpResponseRecord
-impl TryFrom<&reqwest::Response> for HttpResponseRecord {
-    type Error = HttpRecordError;
-
-    fn try_from(response: &reqwest::Response) -> Result<Self, Self::Error> {
-        let (status_code, status_string) = status(&response.status());
-
-        Ok(Self {
-            version: version(&response.version()),
-            status_code,
-            status_string,
-            headers: Some(headers(response.headers())),
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -113,7 +67,7 @@ mod tests {
     #[allow(unused_imports)]
     use rstest_reuse::{self, *};
 
-    fn data_test_header(
+    pub fn data_test_header(
         k_prefix: &str,
         v_prefix: &str,
         count_header: usize,
@@ -151,19 +105,6 @@ mod tests {
             expect_output_vec_type,
             expect_output_vec_str.join("\n"),
         )
-    }
-
-    #[rstest(fuzz_input, case("basic"))]
-    fn test_valid_format_reqwest_one_header(fuzz_input: &str) {
-        let (test_header_map, expected_result, _) = data_test_header(fuzz_input, fuzz_input, 1);
-        assert_eq!(headers(&test_header_map), expected_result);
-    }
-    // Unicorns not welcome as headers :(
-    #[rstest(fuzz_input, case("🦄"))]
-    #[should_panic]
-    fn test_invalid_format_reqwest_one_header(fuzz_input: &str) {
-        let (test_header_map, expected_result, _) = data_test_header(fuzz_input, fuzz_input, 1);
-        assert_eq!(headers(&test_header_map), expected_result);
     }
 
     #[rstest(fuzz_body_input, case("basic"), case("🦄"))]
